[qemu-utils] Move qemu utils out of qemu backend
