+++ include/multipass/utils/qemu_img_utils.h
