+++ src/daemon/default_vm_image_vault.cpp
@@ -31,8 +31,8 @@
 #include <multipass/rpc/multipass.grpc.pb.h>
 #include <multipass/url_downloader.h>
 #include <multipass/utils.h>
+#include <multipass/utils/qemu_img_utils.h>
 #include <multipass/vm_image.h>
-#include <qemu/qemu_img_utils.h>
 
 #include <QUrl>
 #include <QtConcurrent/QtConcurrent>
