+++ src/platform/backends/qemu/CMakeLists.txt
@@ -18,7 +18,6 @@ add_definitions(-DHOST_ARCH="${HOST_ARCH}")
 
 add_library(qemu_backend STATIC
   qemu_base_process_spec.cpp
-  qemu_img_utils.cpp
   qemu_mount_handler.cpp
   qemu_snapshot.cpp
   qemu_vm_process_spec.cpp
