+++ src/platform/backends/qemu/qemu_virtual_machine_factory.cpp
@@ -16,7 +16,6 @@
  */
 
 #include "qemu_virtual_machine_factory.h"
-#include "qemu_img_utils.h"
 #include "qemu_virtual_machine.h"
 
 #include <multipass/cloud_init_iso.h>
@@ -24,6 +23,7 @@
 #include <multipass/logging/log.h>
 #include <multipass/platform.h>
 #include <multipass/process/simple_process_spec.h>
+#include <multipass/utils/qemu_img_utils.h>
 #include <multipass/yaml_node_utils.h>
 
 #include <QRegularExpression>
